pyproject.toml
[project]
name = "agro-risk-intelligence"
version = "0.1.0"
description = "Agro Market Intelligence: Análise Quantitativa e Monitoramento Estratégico de Risco (PR & MT)."
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "pandas",
    "yfinance",
    "matplotlib",
    "seaborn",
    "python-dotenv",
    "pytz",
    "supabase",
    "requests",  # Essencial para o novo mailer.py
    "streamlit>=1.53.0",
    "plotly>=6.5.2",
    "feedparser>=6.0.12",
    "pyyaml>=6.0.3",
    "numpy==2.4.1",
    "fpdf>=1.7.2",
    "mailersend>=2.0.0",
    "huggingface-hub>=1.3.2",
    "pypdf>=6.6.1",
    "xlrd>=2.0.2",
    "scipy>=1.17.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
